Update #1
Well, this seems like a massive update, but in reality, it's not really. I am finally using an asset pack a I purchased a while ago on a sale! Changed the projectile effects, and set it to manual. Updated map. Figured out my roadmap for the game! Added escaping the map objective.