Update #7 - Probably Final Update
Finished the final room. It isn't what I initially wanted for it, but due to time constraints and having to deal my time a bit different, this is just going to have to be where it is for this project. At some point in the future I'll probably change some things like more mechanics. But, this feels like the most complete game I've made so far, so that's a plus I suppose.